# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": " Movieflix API",
    "description": "API responsavel por gerenciar os filmes do Movieflix",
    "version": "5.0.1"
  },
  "paths": {
    "/movies": {
      "get": {
        "tags": ["Movie"],
        "summary": "Obter todos os filmes",
        "responses": {
          "200": {
            "description": "Retorna uma lista de filmes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Movie"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/movies/{id}": {
      "put": {
        "tags": ["Movie"],
        "summary": "Atualizar um filme",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID do filme a ser atualizado",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MovieInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Filme atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/MovieInput"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Filme não encontrado"
          },
          "500": {
            "description": "Falha ao atualizar o filme"
          }
        }
      },
      "delete": {
        "tags": ["Movie"],
        "summary": "Remover um filme",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID do filme a ser removido",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Filme removido com sucesso"
          },
          "404": {
            "description": "Filme não encontrado"
          },
          "500": {
            "description": "Não foi possivel remover o filme"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Movie": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "Titulo do filme"
          },
          "genre_id": {
            "type": "integer",
            "description": "Genero do filme"
          },
          "genres": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "description": "ID do genero"
              },
              "name": {
                "type": "string",
                "description": "Nome do genero"
              }
            }
          },
          "language_id": {
            "type": "integer",
            "description": "Idioma do filme"
          },
          "languages": {
            "type": "object",
            "properties": {
              "id": {
                "type": "integer",
                "description": "ID do genero"
              },
              "name": {
                "type": "string",
                "description": "Nome do Idioma"
              }
            }
          },
          "oscar_count": {
            "type": "integer",
            "description": "Quantidade de oscar do filme"
          },
          "release_date": {
            "type": "string",
            "format": "date",
            "description": "Data de lançamento do filme"
          }
        }
      },
      "MovieInput": {
        "type": "object",
        "required": [
          "title",
          "genre_id",
          "language_id",
          "oscar_count",
          "release_date"
        ],

        "properties": {
          "title": {
            "type": "string",
            "description": "Titulo do filme"
          },
          "genre_id": {
            "type": "integer",
            "description": "ID do genero do filme"
          },
          "language_id": {
            "type": "integer",
            "description": "ID da Lingua do filme"
          },
          "oscar_count": {
            "type": "integer",
            "description": "Quantidade de oscar que o filme recebeu"
          },
          "release_date": {
            "type": "string",
            "formt": "date",
            "description": "Data do lançamento do filme"
          }
        }
      }
    }
  }
}